+++ src/test/java/com/ubiqsecurity/UbiqConfigurationTest.java
@@ -33,7 +33,8 @@ public class UbiqConfigurationTest
       assertEquals(cfg.getEventReportingFlushInterval().compareTo(10),0);
       assertEquals(cfg.getEventReportingTrapExceptions(), false);
       assertEquals(cfg.getKeyCacheEncryptKeys(), false);
-      assertEquals(cfg.getKeyCacheUnstructuredKeys(), false);
+      assertEquals(cfg.getKeyCacheUnstructuredKeys(), true);
+      assertEquals(cfg.getKeyCacheStructuredKeys(), true);
       assertEquals(cfg.getKeyCacheTtlSeconds().compareTo(1800), 0);
     }
 
@@ -93,7 +94,8 @@ public class UbiqConfigurationTest
       myWriter.write("\"key_caching\" : { ");
       myWriter.write("\"ttl_seconds\" : 2400,");
       myWriter.write("\"encrypt\" : true,");
-      myWriter.write("\"unstructured\" : true");
+      myWriter.write("\"unstructured\" : false,");
+      myWriter.write("\"structured\" : false");
       myWriter.write("}}");
       myWriter.close();
 
@@ -105,7 +107,8 @@ public class UbiqConfigurationTest
       assertEquals(cfg.getEventReportingFlushInterval().compareTo(390),0);
       assertEquals(cfg.getEventReportingTrapExceptions(), true);
       assertEquals(cfg.getKeyCacheEncryptKeys(), true);
-      assertEquals(cfg.getKeyCacheUnstructuredKeys(), true);
+      assertEquals(cfg.getKeyCacheUnstructuredKeys(), false);
+      assertEquals(cfg.getKeyCacheStructuredKeys(), false);
       assertEquals(cfg.getKeyCacheTtlSeconds().compareTo(2400), 0);
       file.deleteOnExit();
      
@@ -169,8 +172,9 @@ public class UbiqConfigurationTest
       assertEquals(cfg.getEventReportingFlushInterval().compareTo(defaultCfg.getEventReportingFlushInterval()),0);
       assertEquals(cfg.getEventReportingTrapExceptions(), defaultCfg.getEventReportingTrapExceptions());
 
-      assertEquals(cfg.getKeyCacheEncryptKeys(), false);
-      assertEquals(cfg.getKeyCacheUnstructuredKeys(), false);
+      assertEquals(cfg.getKeyCacheEncryptKeys(), defaultCfg.getKeyCacheEncryptKeys());
+      assertEquals(cfg.getKeyCacheUnstructuredKeys(), defaultCfg.getKeyCacheUnstructuredKeys());
+      assertEquals(cfg.getKeyCacheStructuredKeys(), defaultCfg.getKeyCacheStructuredKeys());
       assertEquals(cfg.getKeyCacheTtlSeconds().compareTo(2400), 0);